requirements.txt
numpy
pandas
msgpack-python
spacy
jsonlines
